test: split tests in gridViewIT
